# appstore/settings/openapi.yaml
openapi: 3.0.0
info:
  version: "1.3.1"
  title: IDS Metadata Broker API
  description: This is the reference API description for IDS Metadata Broker HTTP Calls. Note that this is, in general, not a REST API.
  contact:
    email: contact@ids.fraunhofer.de
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
    
# servers:
# - url: https://{authority}/{basePath}
#   description: The IDS Metadata Broker can be deployed in any domain. As usualy the broker is not the only resource served by one domain, either subdomains (e.g. 'broker.ids.isst.fraunhofer.de') or paths ('example.org/path/to/the/broker/') need to be specified. This API description treats the trailing slash ('/') as the root location of its paths.
#   variables:
#     authority:
#       default: broker.ids.isst.fraunhofer.de
#       description: The authority of the server URL containing the domain name and port.
#     basePath:
#       default: ' '
#       description: The base path at the server to the Broker endpoint.
# - description: Example URL
#   url: https://example.com/ids/broker/
...
paths:
...
  /api/apps:
    get:
      tags:
        - AppStore
      summary: TODO
      description: TODO
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  schema:
                    $ref: '#/components/schemas/AppMetadata'

    post:
      tags:
        - AppStore
      summary: TODO
      description: TODO
      requestBody:
        description: TODO body description
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/AppMetadata'
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AppMetadata'
  /api/apps/search:
    get:
      tags:
        - AppStore
      summary: todo
      description: todo
      parameters:
        - name: q
          in: query
          type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  schema:
                    $ref: '#/components/schemas/AppMetadata'
  /api/apps/{appID}/metadata:
    get:  
      tags:
        - AppStore
      summary: todo
      description: todo
      parameters:
        - name: appID
          in: path
          required: true
          type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AppMetadata'
    put:  
      tags:
        - AppStore
      summary: todo
      description: todo
      parameters:
        - name: appID
          in: path
          required: true
          type: string
      requestBody:
        description: TODO body description
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/AppMetadata'
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AppMetadata'
...
components:
  schemas:
    AppMetadata:
      type: object
      properties:
        name: 
          type: string
          example: Example nginx
        introduction: 
          type: string
          example: Short text introducing my nginx App
        description:
          type: string
          example: <p>A long, <b>HTML</b> formatted description explaining how to use the App (like a user manual).</p>
        keywords:
          type: array
          items: 
            type: string
          example: [ 'nginx', 'proxy' ]
...